Experimental (#59)
* Say Command
Fixes #34

* Purge Command
Fixes #49

* Auto Role On Invite Link
Success Setup

* info update

* Fixed Avatarless User Authenticating
Fixes #53

* v4.3.6

* Auto Role On Invite Link
Fixes #48

* v4.4.0 Stable Update

* v4.4.0 Stable Update

+++ simp-bot/src/commands/admin/SayCommand.js
@@ -1,5 +1,10 @@
 const BaseCommand = require('../../utils/structures/BaseCommand');
 
+/**
+ * Allows server admin to have the bot send a message in a specified channel
+ * @version 4.3.3
+ */
+
 module.exports = class SayCommand extends BaseCommand {
   constructor() {
     super('say', 'admin', []);
@@ -7,11 +12,32 @@ module.exports = class SayCommand extends BaseCommand {
 
   async run(client, message, args) {
 
-    await message.channel.bulkDelete(1);
-    var newMessage = "";
-    for (var i in args){
-      newMessage += ' '+args[i];
+    var Permissions = message.member.permissions;
+    if(!Permissions.has('ADMINISTRATOR')) {
+      message.channel.bulkDelete(1);
+      message.reply("you dont have permissions to complete this action.")
+      .then(message => {
+        message.delete({ timeout: 5000});
+      })
+      .catch(err => {
+        throw err
+      });
+      return;
     }
-    message.channel.send(newMessage);
+
+    try {
+      let channel = args[0].slice(2,args[0].length-1);
+      let botMessage = args.slice(1).join(" ");
+      
+
+      client.channels.cache.get(channel).send(botMessage);
+    }
+    catch (err) {
+      await message.channel.bulkDelete(1);
+      let newMessage = args.slice(0).join(" ");
+
+      message.channel.send(newMessage);
+    }
+    
   }
 }
